pyproject.toml
[tool.poetry]
name = "vendoj-auth-service"
version = "0.1.0"
description = "Authentication and authorization service for Vendoj"
authors = ["igortxra <igor.teixeira@cromai.com>"]
readme = "README.md"
packages = [{include = "vendoj_auth_service"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
gunicorn = "^20.1.0"
bcrypt = "^4.0.1"
pyjwt = "^2.6.0"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
flask-cli = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
